Closes #93: Extract shared code across services and controllers
- Extract RecipeImport::Normalizer module from identical normalize methods
  in AiExtractor and PhotoExtractor
- Extract RecipeParamsPermittable concern from duplicated recipe_params
  in web and API recipe controllers
- Extract MealPlanSetupable concern from duplicated generate_days and
  attach_participants in web and API meal plan controllers
