+++ app/controllers/accounts/api/v1/recipes_controller.rb
@@ -1,5 +1,6 @@
 class Accounts::Api::V1::RecipesController < Accounts::Api::V1::ApplicationController
   include AiRateLimited
+  include RecipeParamsPermittable
 
   before_action :require_write_permission!, only: %i[create update destroy import_url import_photo import_confirm calculate_nutrition]
   before_action :set_recipe, only: %i[show update destroy calculate_nutrition]
@@ -218,22 +219,6 @@ class Accounts::Api::V1::RecipesController < Accounts::Api::V1::ApplicationContr
   end
 
   def recipe_params
-    params.require(:recipe).permit(
-      :title,
-      :description,
-      :category,
-      :source,
-      :servings,
-      :prep_time,
-      :cook_time,
-      :rating,
-      ingredients_attributes: %i[id name quantity unit display_order _destroy],
-      instructions_attributes: %i[id step_number instruction _destroy],
-      nutrition_data_attributes: %i[
-        id calories fat protein carbs fiber
-        sodium potassium magnesium _destroy
-      ],
-      tips_attributes: %i[id tip _destroy]
-    )
+    permitted_recipe_params
   end
 end
