+++ app/controllers/accounts/meal_plans_controller.rb
@@ -1,5 +1,6 @@
 class Accounts::MealPlansController < ApplicationController
   include AiRateLimited
+  include MealPlanSetupable
 
   before_action :set_meal_plan, only: %i[show edit update destroy duplicate regenerate_day]
   before_action -> { check_ai_rate_limit!(:meal_plan_generation, redirect_path: new_meal_plan_path) },
@@ -164,23 +165,6 @@ class Accounts::MealPlansController < ApplicationController
 
   def meal_plan_params
     params.require(:meal_plan).permit(:name, :start_date, :end_date, :daily_calories_target)
-  end
-
-  def generate_days(plan)
-    (plan.start_date..plan.end_date).each_with_index do |date, index|
-      plan.days.create!(date: date, day_number: index + 1)
-    end
-  end
-
-  def attach_participants(plan, profile_ids)
-    return unless profile_ids.present?
-
-    Array(profile_ids).reject(&:blank?).each do |profile_id|
-      profile = Current.account.dietary_profiles.active.find_by(id: profile_id)
-      next unless profile
-
-      plan.participants.create!(dietary_profile: profile)
-    end
   end
 
   def sync_days(plan, old_start, old_end)
