+++ app/services/recipe_import/ai_extractor.rb
@@ -2,6 +2,8 @@
 
 module RecipeImport
   class AiExtractor
+    include RecipeImport::Normalizer
+
     SYSTEM_PROMPT = <<~PROMPT
       You are a recipe extraction assistant. Given the text content of a web page,
       extract the recipe information and return it using the provided tool.
@@ -59,33 +61,6 @@ module RecipeImport
       )
 
       normalize(result[:input])
-    end
-
-    private
-
-    def normalize(input)
-      instructions = (input["instructions"] || []).map.with_index(1) do |text, i|
-        { step_number: i, instruction: text }
-      end
-
-      nutrition = {
-        calories: input["calories"],
-        fat: input["fat"],
-        protein: input["protein"],
-        carbs: input["carbs"],
-        fiber: input["fiber"]
-      }.compact.presence
-
-      {
-        title: input["title"],
-        description: input["description"],
-        servings: input["servings"],
-        prep_time: input["prep_time"],
-        cook_time: input["cook_time"],
-        ingredients: input["ingredients"] || [],
-        instructions: instructions,
-        nutrition: nutrition
-      }.compact
     end
   end
 end
